# packages/backend/openapi.yml
openapi: 3.0.3
info:
  title: Bid Eligibility Judgment System API
  version: 0.1.0
  description: Backend API for bid evaluation, announcements, partners, orderers, contacts, and companies.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://bidapp-backend-postgres-50843898931.asia-northeast1.run.app
    description: Cloud Run production
paths:
...
  /api/orderers/{id}:
    get:
      summary: Get orderer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Single orderer
        '404':
          description: Not found
...